Enable detekt linter and analysis
