+++ gradle-plugins/build.gradle.kts
@@ -1,8 +1,10 @@
 repositories {
     gradlePluginPortal()
+    mavenCentral()
 }
 
 val kotlinVersion = "2.0.0-RC1"
+val detektVersion = "1.23.6"
 
 plugins {
     `kotlin-dsl`
@@ -14,4 +16,5 @@ version = "0.0.1-SNAPSHOT"
 
 dependencies {
     implementation("org.jetbrains.kotlin.jvm:org.jetbrains.kotlin.jvm.gradle.plugin:$kotlinVersion")
+    implementation("io.gitlab.arturbosch.detekt:detekt-gradle-plugin:$detektVersion")
 }
