--- server/src/main/kotlin/Main.kt
@@ -1,137 +1,0 @@
-package io.github.akmal2409.nats.server
-
-import io.github.akmal2409.nats.server.parser.ParsingError
-import io.github.akmal2409.nats.server.parser.PendingParsing
-import io.github.akmal2409.nats.server.parser.SuspendableParser
-import io.github.akmal2409.nats.transport.ClientRequest
-import io.github.akmal2409.nats.transport.ConnectionHandler
-import io.github.akmal2409.nats.transport.DeserializationResult
-import io.github.akmal2409.nats.transport.NioTcpServerSocketChannelFactory
-import io.github.akmal2409.nats.transport.RequestDeserializer
-import io.github.akmal2409.nats.transport.SelectorFactoryImpl
-import io.github.akmal2409.nats.transport.Transport
-import io.github.akmal2409.nats.transport.common.scopedFlow
-import io.github.oshai.kotlinlogging.KotlinLogging
-import java.net.InetSocketAddress
-import java.nio.ByteBuffer
-import java.time.Instant
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.catch
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.onEach
-import kotlinx.coroutines.flow.produceIn
-import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.selects.select
-
-class ScheduledPing(
-    val timeoutReached: Boolean = false
-)
-
-object Handler : ConnectionHandler<Request, ByteBuffer> {
-
-    private val logger = KotlinLogging.logger {}
-
-    private val scope = CoroutineScope(Dispatchers.Default)
-
-    override fun onConnect(request: ClientRequest<Request>): Flow<ByteBuffer> {
-        logger.info { "Received request in handler ${request.remoteAddress}" }
-
-        return scopedFlow<ByteBuffer> { requestScope ->
-            val channel = request.requestFlow.produceIn(requestScope)
-            val pingChannel = pingFlow().produceIn(requestScope)
-
-            while (true) {
-
-                select {
-                    channel.onReceive {
-                        logger.info { "Received request $request" }
-
-                        emit(convertToResponse(PingResponse))
-                        emit(mockText())
-                    }
-
-//                    pingChannel.onReceive {
-//                        if (it.timeoutReached) {
-//                            logger.info { "Timeout reached for client. Terminating connection" }
-//                            throw RuntimeException("Timeout reached")
-//                        } else {
-//                            logger.info { "Pinging client" }
-//                            emit(convertToResponse(PingResponse))
-//                        }
-//                    }
-                }
-
-
-            }
-        }.catch { ex ->
-            logger.error(ex) { "Error in flow" }
-            throw ex
-        }.onEach { logger.info { "Emitting response $it" } }
-    }
-
-    private fun mockText(): ByteBuffer {
-        val someText = buildString {
-            repeat(10000) {
-                appendLine("THIS is a very long text that will span many many lines")
-            }
-        }
-        return ByteBuffer.wrap("""
-            $someText
-            Hey there pretty
-        """.trimIndent().toByteArray(Charsets.UTF_32))
-    }
-
-
-    private fun pingFlow() = flow<ScheduledPing> {
-        var lastPing: Instant = Instant.EPOCH
-
-        while (true) {
-            val now = Instant.now()
-
-            if (lastPing != Instant.EPOCH && now.isAfter(lastPing.plusSeconds(4))) {
-                emit(ScheduledPing(timeoutReached = true))
-                delay(12000)
-            } else {
-                lastPing = Instant.now()
-                emit(ScheduledPing())
-                delay(5000)
-            }
-        }
-    }
-}
-
-class Deser : RequestDeserializer<Request> {
-
-    private val parser = SuspendableParser(10 * 1024 * 1024, 1024)
-
-    override fun deserialize(bytes: ByteBuffer): DeserializationResult<Request> {
-        return when (val result = parser.tryParse(bytes)) {
-            is PendingParsing -> DeserializationResult.of(null)
-            is ParsingError -> error("Parsing error $result")
-            else -> DeserializationResult.of(convertToRequest(result))
-        }
-    }
-
-    override fun close() {
-        parser.close()
-    }
-}
-
-fun main() = runBlocking {
-    val server = Transport(
-        serverSocketChannelFactory = NioTcpServerSocketChannelFactory(
-            InetSocketAddress("localhost", 5000)
-        ),
-        selectorFactory = SelectorFactoryImpl(),
-        connectionHandler = Handler,
-        requestDeserializerFactory = { Deser() },
-        responseSerializer = { it }
-    )
-
-    server.start()
-
-    Runtime.getRuntime().addShutdownHook(Thread(server::stop))
-}
