--- server/src/main/kotlin/RequestConverters.kt
@@ -1,4 +1,0 @@
-package io.github.akmal2409.nats.server
-
-class RequestConverters {
-}
