+++ server/src/main/kotlin/io/github/akmal2409/knats/server/parser/ConnectOptions.kt
@@ -1,0 +1,12 @@
+package io.github.akmal2409.knats.server.parser
+
+data class ConnectOptions(
+    val verbose: Boolean = false, // turns on +OK
+    val pedantic: Boolean = false,
+    val tlsRequired: Boolean = false,
+) {
+
+    fun toJson(): String = """
+        {"verbose": $verbose, "pedantic": $pedantic, "tlsRequired": $tlsRequired}
+    """.trimIndent()
+}
